Move "Network::finalUrl" to "HTTPRequest" class
